pyproject.toml
[tool.poetry]
name = "proof-of-concept"
version = "0.1.0"
description = ""
authors = ["cucupac <46691282+cucupac@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.7.2"
black = "^24.4.2"
web3 = "^6.19.0"
python-dotenv = "^1.0.1"
numpy = "^1.26.4"
eth-abi = "^5.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
